refactor(chat): extract model-resolution helpers from chat-context
chat-context.tsx (1552 lines) mixed provider components with three
pure, self-contained functions (findModel, resolveActiveTier,
pickFallbackChatModel) that only read their own args. Moved them
to resolve-chat-model.ts unchanged, just relocated + imported.
